libdl: Add powerpc large memory and small data support.
- Add support for architecure sections that can be handled by the
  architecture back end.

- Add trampoline/fixup support for PowerPC. This means the PowerPC
  now supports large memory loading of applications.

- Add a bit allocator to manage small block based regions of memory.

- Add small data (sdata/sbss) support for the PowerPC. The support
  makes the linker allocated small data region of memory a global
  resource available to libdl loaded object files.

Updates #3687
Updates #3685

+++ cpukit/libdl/rtl-shell.c
@@ -175,7 +175,7 @@ rtems_rtl_dependencies (rtems_rtl_obj* obj,
 {
   rtems_rtl_dep_data* dd = (rtems_rtl_dep_data*) data;
   if (!dd->first)
-    printf ("\n%-*c: ", dd->indent, ' ');
+    printf ("\n%-*c: ", (int) dd->indent, ' ');
   else
     dd->first = false;
   printf ("%s", dependent->oname);
